added options to keep intermediate install files and to skip redownloading them

+++ src/rna/driver/rna_installer.py
@@ -60,7 +60,9 @@ class RailRnaInstaller(object):
 
     def __init__(self, zip_name, curl_exe=None, install_dir=None,
                     no_dependencies=False, prep_dependencies=False,
-                    yes=False, me=False, add_symlinks=False):
+                    yes=False, me=False, add_symlinks=False, 
+                    keep_intermediate_files=False,
+                    previous_temp_install_dir=None):
         print_to_screen(u"""{0} Rail-RNA v{1} Installer""".format(
                                         u'\u2200', version_number)
                                     )
@@ -87,11 +89,18 @@ class RailRnaInstaller(object):
         self.log_file = os.path.join(log_dir, 'rail-rna_install.log')
         self.log_stream = open(self.log_file, 'w')
         self.finished = False
+        self.keep_intermediate_files = keep_intermediate_files
         register_cleanup(remove_temporary_directories, [log_dir])
         self.yes = yes or me
         self.me = me
         self.prep_dependencies = prep_dependencies
         self.add_symlinks = add_symlinks
+        self.previous_temp_install_dir = previous_temp_install_dir
+        # Install in a temporary directory first, then move to final dest
+        if self.previous_temp_install_dir:
+            self.temp_install_dir = self.previous_temp_install_dir
+        else:
+            self.temp_install_dir = tempfile.mkdtemp()
 
     def __enter__(self):
         return self
@@ -190,13 +199,22 @@ class RailRnaInstaller(object):
         url_deque = deque(urls)
         while url_deque:
             url = url_deque.popleft()
+            download_file = url.split('/').pop()
             command = [self.curl_exe, '-L', '-O', url]
             filename = url.rpartition('/')[2]
-            print >>self.log_stream, 'Downloading {} from {}...'.format(
-                                                                    name, url
-                                                                )
             try:
-                subprocess.check_output(command, stderr=self.log_stream)
+                if (self.previous_temp_install_dir and 
+                    os.path.exists(os.path.join(
+                                   self.previous_temp_install_dir, 
+                                   download_file))):
+                    print >>self.log_stream, ('skipping '
+                              'download of {0} since it\'s already '
+                              'present in {1}'.format(download_file, 
+                                            self.previous_temp_install_dir))
+                else:
+                    print >>self.log_stream, ('Downloading '
+                              '{} from {}...'.format(name, url))
+                    subprocess.check_output(command, stderr=self.log_stream)
             except subprocess.CalledProcessError as e:
                 if not url_deque:
                     self._print_to_screen_and_log(
@@ -246,7 +264,8 @@ class RailRnaInstaller(object):
                             )
                             continue
                     finally:
-                        os.remove(filename)
+                        if not self.keep_intermediate_files:
+                            os.remove(filename)
                 if explode_command is not None:
                     self._print_to_screen_and_log(
                             '[Installing] Extracting %s...' % name,
@@ -273,7 +292,8 @@ class RailRnaInstaller(object):
                             )
                             continue
                     finally:
-                        os.remove(filename)
+                        if not self.keep_intermediate_files:
+                            os.remove(filename)
                 break
 
     def _quote(self, path=None):
@@ -322,9 +342,9 @@ class RailRnaInstaller(object):
         else:
             # User specified an installation directory
             self.final_install_dir = self.install_dir
-        # Install in a temporary directory first, then move to final dest
-        temp_install_dir = tempfile.mkdtemp()
-        register_cleanup(remove_temporary_directories, [temp_install_dir])
+        if not self.keep_intermediate_files:
+            #this will also get ipython and aws temp install directories
+            register_cleanup(remove_temporary_directories, [self.temp_install_dir])
         if os.path.exists(self.final_install_dir):
             if self._yes_no_query(
                     ('The installation path {dir} already exists.\n    '
@@ -361,12 +381,16 @@ class RailRnaInstaller(object):
             # So it's possible to move temp installation dir there
             os.rmdir(self.final_install_dir)
             pass
-        with cd(temp_install_dir):
-            with zipfile.ZipFile(self.zip_name) as zip_object:
-                zip_object.extractall('./rail-rna')
+        with cd(self.temp_install_dir):
+            if not (self.previous_temp_install_dir and os.path.exists("./rail-rna")):
+                with zipfile.ZipFile(self.zip_name) as zip_object:
+                    zip_object.extractall('./rail-rna')
             if not self.no_dependencies:
                 self._grab_and_explode(self.depends['pypy'], 'PyPy')
                 if not self.prep_dependencies:
+                    #_grab_and_explode will check to see if these are 
+                    #already downloaded and if previous_temp_install_dir is set 
+                    #will avoid redownloading them
                     self._grab_and_explode(self.depends['bowtie1'], 'Bowtie 1')
                     self._grab_and_explode(self.depends['bowtie2'], 'Bowtie 2')
                     self._grab_and_explode(self.depends['bedgraphtobigwig'],
@@ -375,7 +399,7 @@ class RailRnaInstaller(object):
                                                                 'SAMTools')
             if not self.prep_dependencies:
                 # Have to make SAMTools (annoying; maybe change this)
-                samtools_dir = os.path.join(temp_install_dir,
+                samtools_dir = os.path.join(self.temp_install_dir,
                         self.depends['samtools'][0].rpartition('/')[2][:-8]
                     )
                 with cd(samtools_dir):
@@ -447,7 +471,7 @@ class RailRnaInstaller(object):
                 )
             # Write paths to exe_paths
             with open(
-                            os.path.join(temp_install_dir, 'rail-rna',
+                            os.path.join(self.temp_install_dir, 'rail-rna',
                                             'exe_paths.py'), 'w'
                         ) as exe_paths_stream:
                 print >>exe_paths_stream, (
@@ -478,13 +502,14 @@ bedgraphtobigwig = {bedgraphtobigwig}
                             bedgraphtobigwig=self._quote(bedgraphtobigwig))
         # Move to final directory
         try:
-            shutil.move(temp_install_dir, self.final_install_dir)
+            #shutil.move(self.temp_install_dir, self.final_install_dir)
+            shutil.copytree(self.temp_install_dir, self.final_install_dir)
         except Exception as e:
             self._print_to_screen_and_log(('Problem "%s" encountered moving '
                                            'temporary installation directory '
                                            '%s to final destination %s.') % (
                                                 e,
-                                                temp_install_dir,
+                                                self.temp_install_dir,
                                                 self.final_install_dir
                                             ))
             self._bail()
@@ -616,9 +641,9 @@ fi
                     'work in its "parallel" mode.\n'
                     '    * Install IPython now?'
                 ):
-                temp_ipython_install_dir = tempfile.mkdtemp()
-                register_cleanup(remove_temporary_directories,
-                                    [temp_ipython_install_dir])
+                #temp_ipython_install_dir = tempfile.mkdtemp()
+                temp_ipython_install_dir = self.temp_install_dir
+                
                 with cd(temp_ipython_install_dir):
                     self._grab_and_explode(self.depends['ipython'], 'IPython')
                     setup_dir = os.path.dirname(find('setup.py', './'))
@@ -645,9 +670,7 @@ fi
                 'in its "elastic" mode, on Amazon Elastic MapReduce.\n'
                 '    * Install AWS CLI now?'
             ):
-            temp_aws_install_dir = tempfile.mkdtemp()
-            register_cleanup(remove_temporary_directories,
-                                [temp_aws_install_dir])
+            temp_aws_install_dir = self.temp_install_dir
             with cd(temp_aws_install_dir):
                 self._grab_and_explode(self.depends['aws_cli'], 'AWS CLI')
                 os.chmod('./awscli-bundle/install', 0755)
